struct(T7): split world.rs into scene/charts/deals/venues package
Pure move of GameWorld methods and types into world/{mod,scene,charts,
deals,venues}.rs. update_week stays the conductor in mod.rs; public types
and scene/chart constants re-export from the package root. World unit
tests remain under world::tests with stable names. Disjoint from T2.
